requirements.txt
bertopic==0.15.0
gensim==4.3.2
hdbscan==0.8.33
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
pandas==1.3.4
pandas_stubs==1.2.0.35
scikit_learn==1.2.2
sentence_transformers==3.4.1
spacy==3.7.1
umap-learn==0.5.3
wordcloud==1.9.2
keybert==0.4.0
openpyxl==3.0.10
plotly==5.9.0
